+++ src/pages/ContractManagement/PaymentGateway/payment-details.tsx
